app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
passlib[bcrypt]
python-jose[cryptography]
slowapi
email-validator
bleach